# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 回测平台 API
  version: 1.0.0
  description: |
    回测平台提供组合回测、资产分析、蒙特卡洛模拟、组合优化、有效前沿计算、
    战术分配、信号分析、PCA 主成分分析、LETF 滑点分析、目标优化等量化投资工具。

    ## 认证
    - 管理端点（admin、data/manage）需要 `x-api-key` 请求头认证
    - 计算密集型端点（backtest、backtest-optimizer）可选 `x-api-key`，有 Key 时验证身份，无 Key 时放行但记录匿名访问
    - 健康检查和 metrics 端点不需要认证

    ## 速率限制
    - 普通 API：100 次/15 分钟/IP
    - 计算密集型 API（backtest、backtest-optimizer）：10 次/分钟/IP
servers:
  - url: http://localhost:5001/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌，通过 /auth/login 获取
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: |
        幂等性 Key，用于防止管理端点写操作的重复执行。
        - 客户端在 POST 请求中携带唯一的 Idempotency-Key
        - 相同 Key 的请求在 1 小时内返回缓存结果，不重复执行业务逻辑
        - 仅对 POST 写操作生效，GET 请求忽略此头
        - Key 长度不能超过 128 字符
        - 建议使用 UUID v4 格式
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          oneOf:
            - type: string
            - type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                  description: 机器可读错误码
                message:
                  type: string
                  description: 人类可读错误描述

    # ===== 认证鉴权 =====
    # 企业理由：API 契约文档是前后端协作的唯一可靠依据。
    # 认证流程涉及令牌签发、刷新、撤销，任何字段歧义都会导致集成故障，
    # 因此必须将 LoginRequest / TokenResponse / UserInfo 等结构固化到 OpenAPI 规范中，
    # 确保客户端与服务端对令牌生命周期管理达成一致理解。
...
    Statistics:
      type: object
      properties:
        cagr:
          type: number
        mwrr:
          type: number
        totalReturn:
          type: number
        bestYear:
          type: number
        worstYear:
          type: number
        avgYear:
          type: number
        stdev:
          type: number
        maxDrawdown:
          type: number
        maxDrawdownDuration:
          type: number
        sharpe:
          type: number
        sortino:
          type: number
        calmar:
          type: number
        ulcerIndex:
          type: number
        alpha:
          type: number
        beta:
          type: number
        rSquared:
          type: number
        trackingError:
          type: number
        informationRatio:
          type: number
        upsideCapture:
          type: number
        downsideCapture:
          type: number
        var5:
          type: number
        cvar5:
          type: number
        skewness:
          type: number
        excessKurtosis:
          type: number
        pctPositiveDays:
          type: number
        maxDailyReturn:
          type: number
        minDailyReturn:
          type: number
    DragResult:
      type: object
      properties:
        totalDrag:
          type: number
        annualDrag:
          type: number
        dragSeries:
          type: array
          items:
            type: number
    PortfolioResult:
      type: object
      properties:
        name:
          type: string
        growthCurve:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              value:
                type: number
        drawdownCurve:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              drawdown:
                type: number
        rollingReturns:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              return:
                type: number
        annualReturns:
          type: array
          items:
            type: object
            properties:
              year:
                type: integer
              return:
                type: number
        monthlyReturns:
          type: array
          items:
            type: object
            properties:
              year:
                type: integer
              month:
                type: integer
              return:
                type: number
        statistics:
          $ref: '#/components/schemas/Statistics'
        drag:
          $ref: '#/components/schemas/DragResult'
...
    SignalCondition:
      type: object
      required:
        - indicator
        - period
        - operator
        - threshold
      properties:
        indicator:
          type: string
          enum: [sma, ema, rsi, macd, bollinger, momentum]
        period:
          type: number
        operator:
          type: string
          enum: [gt, lt, cross_above, cross_below]
        threshold:
          type: number
    TradingSignal:
      type: object
      required:
        - id
        - name
        - conditions
        - targetWeights
      properties:
        id:
          type: string
        name:
          type: string
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/SignalCondition'
        targetWeights:
          type: array
          items:
            type: object
            required:
              - ticker
              - weight
            properties:
              ticker:
                type: string
              weight:
                type: number
    TacticalStrategy:
      type: object
      required:
        - id
        - name
        - signals
        - aggregationMethod
      properties:
        id:
          type: string
        name:
          type: string
        signals:
          type: array
          items:
            $ref: '#/components/schemas/TradingSignal'
        aggregationMethod:
          type: string
          enum: [weighted_average, rank, voting]
        rankingConfig:
          type: object
          properties:
            method:
              type: string
              enum: [fixed_share, risk_parity]
            topN:
              type: number
    EmailAlertConfig:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
        email:
          type: string
          format: email
        triggers:
          type: array
          items:
            type: string
            enum: [signal_change, rebalance, threshold]

    # ===== 信号分析 =====
    SignalAnalysisRequest:
      type: object
      required:
        - ticker
        - indicator
        - period
        - threshold
        - startDate
        - endDate
        - signalType
      properties:
        ticker:
          type: string
        indicator:
          type: string
          description: "技术指标：sma, ema, rsi, macd, bollinger"
        period:
          type: number
        threshold:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        signalType:
          type: string
          enum: [entry, exit, both]
    SignalAnalysisResult:
      type: object
      properties:
        signals:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              type:
                type: string
                enum: [buy, sell]
              price:
                type: number
        statistics:
          type: object
          properties:
            totalSignals:
              type: number
            winRate:
              type: number
            avgReturn:
              type: number
            maxDrawdown:
              type: number
            sharpe:
              type: number
        equityCurve:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              value:
                type: number
    DualSignalConfig:
      type: object
      required:
        - signal1
        - signal2
        - combinationMethod
      properties:
        signal1:
          $ref: '#/components/schemas/SignalAnalysisRequest'
        signal2:
          $ref: '#/components/schemas/SignalAnalysisRequest'
        combinationMethod:
          type: string
          enum: [and, or, xor]
    MultiSignalConfig:
      type: object
      required:
        - signals
        - aggregationMethod
      properties:
        signals:
          type: array
          items:
            $ref: '#/components/schemas/SignalAnalysisRequest'
        aggregationMethod:
          type: string
          enum: [weighted, voting, rank]
        weights:
          type: array
          items:
            type: number

    # ===== PCA =====
...
paths:
...
  /tactical/backtest:
    post:
      tags: [tactical]
      summary: 运行战术分配回测
      description: |
        接收战术策略配置，计算信号并运行动态权重回测。
        同时运行等权基准回测作为对比。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strategy
                - startDate
                - endDate
                - startingValue
                - rebalanceFrequency
              properties:
                strategy:
                  $ref: '#/components/schemas/TacticalStrategy'
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                startingValue:
                  type: number
                rebalanceFrequency:
                  type: string
                  enum: [daily, weekly, monthly, quarterly, annual, none]
      responses:
        '200':
          description: 战术回测结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      portfolio:
                        $ref: '#/components/schemas/PortfolioResult'
                      benchmark:
                        $ref: '#/components/schemas/PortfolioResult'
                      signalHistory:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                            activeSignals:
                              type: array
                              items:
                                type: string
                            weights:
                              type: array
                              items:
                                type: object
                                properties:
                                  ticker:
                                    type: string
                                  weight:
                                    type: number
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 回测运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tactical/what-if:
    post:
      tags: [tactical]
      summary: 实时价格查询与信号状态
      description: 接收 ticker 列表，返回实时价格与当前信号状态
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tickers
                - strategy
              properties:
                tickers:
                  type: array
                  items:
                    type: string
                  minItems: 1
                strategy:
                  $ref: '#/components/schemas/TacticalStrategy'
                endDate:
                  type: string
                  format: date
                  description: 可选，默认当天
      responses:
        '200':
          description: 实时信号状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        ticker:
                          type: string
                        currentPrice:
                          type: number
                        signalDate:
                          type: string
                        signalType:
                          type: string
                          enum: [buy, sell, hold]
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 查询失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tactical/alerts:
    post:
      tags: [tactical]
      summary: 保存邮件告警配置
      description: 保存邮件告警配置（内存暂存，进程重启后丢失）
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - config
              properties:
                config:
                  $ref: '#/components/schemas/EmailAlertConfig'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      saved:
                        type: boolean
                      config:
                        $ref: '#/components/schemas/EmailAlertConfig'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 保存失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===== 战术网格搜索 =====
...
  /signal/analyze:
    post:
      tags: [signal]
      summary: 单信号分析
      description: |
        对单个资产运行技术指标信号分析，生成买卖信号点、统计指标和权益曲线。
        支持指标：SMA、EMA、RSI、MACD、Bollinger。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalAnalysisRequest'
      responses:
        '200':
          description: 信号分析结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/SignalAnalysisResult'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到价格数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 信号分析失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /signal/dual:
    post:
      tags: [signal]
      summary: 双信号分析
      description: |
        对两个信号进行组合分析，支持 AND/OR/XOR 组合方式。
        返回各信号独立结果、组合结果和逐日对比。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DualSignalConfig'
      responses:
        '200':
          description: 双信号分析结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      signal1:
                        $ref: '#/components/schemas/SignalAnalysisResult'
                      signal2:
                        $ref: '#/components/schemas/SignalAnalysisResult'
                      combined:
                        $ref: '#/components/schemas/SignalAnalysisResult'
                      comparison:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                            signal1:
                              type: string
                              enum: [buy, sell]
                              nullable: true
                            signal2:
                              type: string
                              enum: [buy, sell]
                              nullable: true
                            combined:
                              type: string
                              enum: [buy, sell]
                              nullable: true
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到价格数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 双信号分析失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /signal/multi:
    post:
      tags: [signal]
      summary: 多信号分析
      description: |
        对多个信号进行聚合分析，支持加权平均、投票、排名三种聚合方式。
        各信号共享同一 ticker 与时间范围。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiSignalConfig'
      responses:
        '200':
          description: 多信号分析结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      aggregated:
                        $ref: '#/components/schemas/SignalAnalysisResult'
                      contributions:
                        type: array
                        items:
                          type: object
                          properties:
                            index:
                              type: number
                            indicator:
                              type: string
                            contribution:
                              type: number
                            statistics:
                              type: object
                              properties:
                                totalSignals:
                                  type: number
                                winRate:
                                  type: number
                                avgReturn:
                                  type: number
                                maxDrawdown:
                                  type: number
                                sharpe:
                                  type: number
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到价格数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 多信号分析失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===== LETF =====
...